loadgamepage, logo and update on leaderboard
Logo is for Arcade game save and exit to main menu, updated on the functionality for that. Also, loadgamepage consists of two button which is load saved free to play and arcade game. As for the leaderboard, I updated on the UI to give it a more consistent look and flow.